PYTHON-5947 Move OTel naming policy out of _telemetry.py into _otel.py
Operation name overrides, the runCommand name and the enum normalization
helper are OpenTelemetry specification rules, so they belong beside
_build_query_summary in _otel.py rather than in the lifecycle module.
Collapse them behind a single _build_operation_name entry point and drop
the unused _OperationTelemetry.operation_name attribute.
